+++ hammurabi/rules/text.py
@@ -19,12 +19,12 @@ from hammurabi.rules.common import SinglePathRule
 class LineExists(SinglePathRule):
     """
     Make sure that the given file contains the required line. This rule is
-    capable for inserting the expected text before or after the unique target
+    capable for inserting the expected text before or after the unique match
     text respecting the indentation of its context.
 
     The default behaviour is to insert the required text exactly after the
-    target line, and respect its indentation. Please note that ``text``and
-    ``target`` parameters are required.
+    match line, and respect its indentation. Please note that ``text``and
+    ``match`` parameters are required.
 
     Example usage:
 
@@ -42,7 +42,7 @@ class LineExists(SinglePathRule):
             >>>             name="Extend gunicorn config",
             >>>             path=gunicorn_config,
             >>>             text="keepalive = 65",
-            >>>             target=r"^bind.*",
+            >>>             match=r"^bind.*",
             >>>             preconditions=[
             >>>                 IsLineNotExists(path=gunicorn_config, criteria=r"^keepalive.*")
             >>>             ]
@@ -55,7 +55,7 @@ class LineExists(SinglePathRule):
 
     .. note::
 
-        The indentation of the target text will be extracted by a simple
+        The indentation of the match text will be extracted by a simple
         regular expression. If a more complex regexp is required, please
         inherit from this class.
     """
@@ -65,14 +65,14 @@ class LineExists(SinglePathRule):
         name: str,
         path: Optional[Path] = None,
         text: Optional[str] = None,
-        target: Optional[str] = None,
+        match: Optional[str] = None,
         position: int = 1,
         respect_indentation: bool = True,
         ensure_trailing_newline: bool = False,
         **kwargs,
     ) -> None:
         self.text = self.validate(text, required=True)
-        self.target = re.compile(self.validate(target, required=True))
+        self.match = re.compile(self.validate(match, required=True))
         self.position = position
         self.respect_indentation = respect_indentation
 
@@ -81,27 +81,27 @@ class LineExists(SinglePathRule):
 
         super().__init__(name, path, **kwargs)
 
-    def __get_target_match(self, lines: List[str]) -> str:
+    def __get_match_match(self, lines: List[str]) -> str:
         """
-        Get the matching target from the content of the given file.
+        Get the matching match from the content of the given file.
         In case the matching number of lines are more than one or no
         match found, an exception will be raised accordingly.
 
         :param lines: Content of the given file
         :type lines: List[str]
 
-        :raises: ``LookupError`` if no matching line can be found for target
+        :raises: ``LookupError`` if no matching line can be found for match
 
         :return: List of the matching line
         :rtype: str
         """
 
-        target_match = list(filter(self.target.match, lines))
+        match_match = list(filter(self.match.match, lines))
 
-        if not target_match:
-            raise LookupError(f'No matching line for "{self.target}"')
+        if not match_match:
+            raise LookupError(f'No matching line for "{self.match}"')
 
-        return target_match.pop()
+        return match_match.pop()
 
     def __get_lines_from_file(self) -> Tuple[List[str], bool]:
         """
@@ -149,16 +149,16 @@ class LineExists(SinglePathRule):
         :type lines: List[str]
         """
 
-        target_match = self.__get_target_match(lines)
+        match_match = self.__get_match_match(lines)
 
         # Get the index of the element from the right
-        target_match_index = len(lines) - lines[::-1].index(target_match) - 1
+        match_match_index = len(lines) - lines[::-1].index(match_match) - 1
 
-        insert_position = target_match_index + self.position
+        insert_position = match_match_index + self.position
 
         logging.debug('Inserting "%s" to position "%d"', self.text, insert_position)
 
-        indentation = self.indentation_pattern.match(lines[target_match_index])
+        indentation = self.indentation_pattern.match(lines[match_match_index])
         if self.respect_indentation and indentation:
             self.text = indentation.group() + self.text
 
@@ -167,7 +167,7 @@ class LineExists(SinglePathRule):
     def task(self) -> Path:
         """
         Make sure that the given file contains the required line. This rule is
-        capable for inserting the expected rule before or after the unique target
+        capable for inserting the expected rule before or after the unique match
         text respecting the indentation of its context.
 
         :raises: ``LookupError``
@@ -340,7 +340,7 @@ class LineReplaced(SinglePathRule):
         :param lines: The new content of the original file
         :type lines: List[str]
 
-        :param match: The matching target in the given file's content
+        :param match: The matching match in the given file's content
         :type match: str
         """